chore(imageprovider): refactor node image provider code out and cleanup dead code (#959)
* factor node image provider code out and cleanup dead code

* remove TODOs and make func private
