+++ pkg/operator/operator.go
@@ -118,9 +118,6 @@ func NewOperator(ctx context.Context, operator *operator.Operator) (context.Cont
 			azurecache.DefaultCleanupInterval),
 	)
 	imageProvider := imagefamily.NewProvider(
-		operator.KubernetesInterface,
-		cache.New(azurecache.KubernetesVersionTTL,
-			azurecache.DefaultCleanupInterval),
 		azClient.ImageVersionsClient,
 		azConfig.Location,
 		azConfig.SubscriptionID,
